config fzf

+++ scripts/shared_function.sh
@@ -158,29 +158,32 @@ function get_now_timestamp() {
 }
 
 #{{{> configure fzf
-function configure_fzf() {
+function configure_fzf_on_linux() {
     # configure in zshrc
-    if test `cat ${HOME}/.zshrc | grep -c "# fzf:FZF_DEFAULT_COMMAND"` = 0
+    if [[ -f ${HOME}/.zshrc ]]
     then
-        echo "" | tee -a ${HOME}/.zshrc > /dev/null
-        echo "# fzf:FZF_DEFAULT_COMMAND" | tee -a ${HOME}/.zshrc > /dev/null
-        echo "if type rg &> /dev/null; then" | tee -a ${HOME}/.zshrc > /dev/null
-        echo "  export FZF_DEFAULT_COMMAND='rg --files'" | tee -a ${HOME}/.zshrc > /dev/null
-        echo "  export FZF_DEFAULT_OPTS='-m'" | tee -a ${HOME}/.zshrc > /dev/null
-        echo "fi" | tee -a ${HOME}/.zshrc > /dev/null
-        echo "" | tee -a ${HOME}/.zshrc > /dev/null
+        if test `cat ${HOME}/.zshrc | grep -c "# fzf:FZF_DEFAULT_COMMAND"` = 0
+        then
+            echo "" | tee -a ${HOME}/.zshrc > /dev/null
+            echo "# fzf:FZF_DEFAULT_COMMAND" | tee -a ${HOME}/.zshrc > /dev/null
+            echo "if type rg &> /dev/null; then" | tee -a ${HOME}/.zshrc > /dev/null
+            echo "  export FZF_DEFAULT_COMMAND='rg --files'" | tee -a ${HOME}/.zshrc > /dev/null
+            echo "  export FZF_DEFAULT_OPTS='-m'" | tee -a ${HOME}/.zshrc > /dev/null
+            echo "fi" | tee -a ${HOME}/.zshrc > /dev/null
+            echo "" | tee -a ${HOME}/.zshrc > /dev/null
+        fi
     fi
 
     # configure in bashrc
-    if test `cat ${HOME}/.bash_profile | grep -c "# fzf:FZF_DEFAULT_COMMAND"` = 0
+    if test `cat ${HOME}/.bashrc | grep -c "# fzf:FZF_DEFAULT_COMMAND"` = 0
     then
-        echo "" | tee -a ${HOME}/.bash_profile > /dev/null
-        echo "# fzf:FZF_DEFAULT_COMMAND" | tee -a ${HOME}/.bash_profile > /dev/null
-        echo "if type rg &> /dev/null; then" | tee -a ${HOME}/.bash_profile > /dev/null
-        echo "  export FZF_DEFAULT_COMMAND='rg --files'" | tee -a ${HOME}/.bash_profile > /dev/null
-        echo "  export FZF_DEFAULT_OPTS='-m'" | tee -a ${HOME}/.bash_profile > /dev/null
-        echo "fi" | tee -a ${HOME}/.bash_profile > /dev/null
-        echo "" | tee -a ${HOME}/.bash_profile > /dev/null
+        echo "" | tee -a ${HOME}/.bashrc > /dev/null
+        echo "# fzf:FZF_DEFAULT_COMMAND" | tee -a ${HOME}/.bashrc > /dev/null
+        echo "if type rg &> /dev/null; then" | tee -a ${HOME}/.bashrc > /dev/null
+        echo "  export FZF_DEFAULT_COMMAND='rg --files'" | tee -a ${HOME}/.bashrc > /dev/null
+        echo "  export FZF_DEFAULT_OPTS='-m'" | tee -a ${HOME}/.bashrc > /dev/null
+        echo "fi" | tee -a ${HOME}/.bashrc > /dev/null
+        echo "" | tee -a ${HOME}/.bashrc > /dev/null
     fi
 }
 #<}}}
